Return the first five news items for a topic
Note, this is a breaking change to the API as we turn `latest_news` into an
array of hashes rather than a single hash, so we'd need to update consumers
of it at the same time.

If we want to include more topics in the filter, then we'll need to add extra
items to `taxon_lookup` which maps the slugs of a browse topic to the content_id
of the equivalent taxon.

The content_id of the taxon can be found by examining the content item using
the content API - so for gov.uk/education, you need to look at gov.uk/api/content/education.